4/pyproject.toml
[project]
name = "4"
version = "0.1.0"
description = ""
authors = [
    {name = "u2er"}
]
readme = "README.md"
requires-python = ">=3.11, <3.14"
dependencies = [
    "pika (>=1.3.2,<2.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "fastapi (>=0.124.4,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "youtube-transcript-api (>=1.2.3,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
